Rich 2

+++ src/main/java/com/example/demo/src/review/model/PostReviewReq.java
@@ -14,7 +14,8 @@ import java.util.List;
 public class PostReviewReq {
     // 유저 id, 별점, 리뷰내용, 리뷰이미지 url
     private int userIdx;
-    private float star;
+//    @NotNull(message = "별점을 입력해주세요.")
+    private Long star;
     private String content;
     private List<String> reviewImgUrl;
